+++ tests/integration/cases/call_transform_no_wrapper/Scala_heterogeneous_strategy_tuple_call@v3.scala
@@ -1,0 +1,6 @@
+object Fixture_call_transform_no_wrapper_Scala_heterogeneous_strategy_tuple_call {
+def process(value: Any = null): Any = null
+process(value = "hello")
+process(value = 42)
+process(value = true)
+}