+++ tests/integration/cases/dict_mixed_scalars/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,7 @@
+object Fixture_dict_mixed_scalars_Scala_heterogeneous_strategy_tuple {
+case class Record0(a: Int, b: String)
+val my_data = Record0(
+    a = 1,
+    b = "x",
+)
+}