+++ tests/integration/cases/empty_dict/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,3 @@
+object Fixture_empty_dict_Scala_heterogeneous_strategy_tuple {
+val my_data = Map()
+}