+++ tests/integration/cases/nested_mixed_types/Scala_heterogeneous_strategy_tuple_sibling@v3.scala
@@ -1,0 +1,6 @@
+object Fixture_nested_mixed_types_Scala_heterogeneous_strategy_tuple_sibling {
+val my_data = List(
+    List[Int](1, 2),
+    List[String]("a", "b"),
+)
+}