+++ tests/integration/cases/tuple_top_level/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,8 @@
+object Fixture_tuple_top_level_Scala_heterogeneous_strategy_tuple {
+val my_data = (
+    1,
+    "email",
+    "a@gmail.com",
+    100,
+)
+}