+++ tests/integration/cases/tuple_triple_record_field/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,11 @@
+object Fixture_tuple_triple_record_field_Scala_heterogeneous_strategy_tuple {
+case class Record0(call: String, args: (Int, String, Boolean))
+val my_data = Record0(
+    call = "send",
+    args = (
+        1,
+        "email",
+        true,
+    ),
+)
+}